tests/artifacts/scoring-blueprints-submission/requirements.txt
any-agent[all]==0.22.0
python-dotenv
fire
